fix: divide description editor and renderer for category creation component
